# api/flea.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "flea.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/flea/admin": {
      "post": {
        "summary": "Admin",
        "operationId": "Admin",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiAdminResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiAdminRequest"
            }
          }
        ],
        "tags": [
          "Flea"
        ]
      }
    },
    "/v1/flea/config": {
      "get": {
        "operationId": "Config",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiConfigResponse"
            }
          }
        },
        "tags": [
          "Flea"
        ]
      }
    },
    "/v1/flea/create_task": {
      "post": {
        "operationId": "CreateTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiTaskDetails"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiTaskDetails"
            }
          }
        ],
        "tags": [
          "Flea"
        ]
      }
    },
    "/v1/flea/healthz": {
      "get": {
        "operationId": "Healthz",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiHealthCheckResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "service",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Flea"
        ]
      }
    },
    "/v1/flea/modify_task/{taskId}": {
      "post": {
        "operationId": "ModifyTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiTaskDetails"
            }
          }
        },
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiModifyTaskRequest"
            }
          }
        ],
        "tags": [
          "Flea"
        ]
      }
    },
    "/v1/flea/start_task/{taskId}": {
      "get": {
        "operationId": "StartTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiStartTaskResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Flea"
        ]
      }
    },
    "/v1/flea/tasks": {
      "get": {
        "operationId": "ListTasks",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiListTasksResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "modelId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "hyperparametersId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "checkpointId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "startTaskId",
            "description": "Tasks are returned lexicographically sorted by taskId\nIf specified only taskIds \u003e= startTaskId will be returned.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "maxItems",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "includeInactive",
            "in": "query",
            "required": false,
            "type": "boolean",
            "format": "boolean"
          }
        ],
        "tags": [
          "Flea"
        ]
      }
    },
    "/v1/flea/tasks/{taskId}": {
      "get": {
        "operationId": "GetTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiTaskDetails"
            }
          }
        },
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Flea"
        ]
      }
    }
  },
  "definitions": {
    "AdminRequestAdminRequestType": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "RELOAD_TOKENS"
      ],
      "default": "UNKNOWN"
    },
    "ConfigResponseBackendType": {
      "type": "string",
      "enum": [
        "INVALID",
        "MEMORY",
        "GOOGLE_CLOUD_STORAGE"
      ],
      "default": "INVALID"
    },
    "HealthCheckResponseServingStatus": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "SERVING",
        "NOT_SERVING"
      ],
      "default": "UNKNOWN"
    },
    "StartTaskResponseRequestStatus": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "REJECTED",
        "APPROVED"
      ],
      "default": "UNKNOWN"
    },
    "apiAdminRequest": {
      "type": "object",
      "properties": {
        "type": {
          "$ref": "#/definitions/AdminRequestAdminRequestType"
        }
      }
    },
    "apiAdminResponse": {
      "type": "object",
      "properties": {
        "info": {
          "type": "string"
        }
      }
    },
    "apiConfigResponse": {
      "type": "object",
      "properties": {
        "backendType": {
          "$ref": "#/definitions/ConfigResponseBackendType"
        }
      }
    },
    "apiHealthCheckResponse": {
      "type": "object",
      "properties": {
        "status": {
          "$ref": "#/definitions/HealthCheckResponseServingStatus"
        }
      }
    },
    "apiListTasksResponse": {
      "type": "object",
      "properties": {
        "startTaskId": {
          "type": "string"
        },
        "maxItems": {
          "type": "integer",
          "format": "int32"
        },
        "taskIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "apiModifyTaskRequest": {
      "type": "object",
      "properties": {
        "taskId": {
          "type": "string"
        },
        "deadline": {
          "type": "string",
          "format": "date-time"
        },
        "active": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
    "apiStartTaskResponse": {
      "type": "object",
      "properties": {
        "status": {
          "$ref": "#/definitions/StartTaskResponseRequestStatus"
        },
        "jobId": {
          "type": "string"
        },
        "uploadTo": {
          "type": "string"
        }
      }
    },
    "apiTaskDetails": {
      "type": "object",
      "properties": {
        "modelId": {
          "type": "string"
        },
        "hyperparametersId": {
          "type": "string"
        },
        "checkpointId": {
          "type": "string"
        },
        "deadline": {
          "type": "string",
          "format": "date-time"
        },
        "taskId": {
          "type": "string"
        },
        "active": {
          "type": "boolean",
          "format": "boolean"
        },
        "link": {
          "type": "string"
        },
        "checkpointLink": {
          "type": "string"
        }
      },
      "title": "This is used by both /create_task, /task/\u003ctaskId\u003e and /modify_task/\u003ctaskId\u003e"
    }
  }
}